jsonrpc: vdscli: migration: fix migrationCreate
In virt migration code, we do logic on the return
value of VM.migrationCreate API, and we expect it
to be a proper response, so with ['status']['code']
field, plus all the extra fields that APIs can add
(see schema for examples).

Current jsonrpc code strips all but the extra fields,
so jsonrpcvdscli has to force fit a success code.

This patch simplify the code by just passing around
a copy of the raw data, leveraging the response module.
This way the code is both simpler and more conformant
to the old XML-RPC vdscli API.

Change-Id: Ie9288d367c519830f6510a2d1b2a5fc1e410949c
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42970
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>