tests: Add @recorded method decorator
Manually recording function arguments and keywords again and again is
not fun and error prone. This patch introduces a new decorator to make
this task easier.

To make a method recordable, mark the method you want to record:

    class Foo:
        @recorded
        def foo(self, a, b, c=3):
            pass
        def bar(self):
            return 42

Methods decorated with @recorded will record all calls to the
instance's __recording__ list:

    f = Foo()
    f.foo(1, 2, 4)
    f.bar()
    f.foo(2, 3)

    print(f.__recording__)
    [("foo", (1, 2), {"c": 4}), ("foo", (2, 3), {"c": 3})]

The new decorator is used now by vmfakelib.Domain. It may be useful for
other fake objects as well.

Change-Id: I43c7750ae1deb5fa9bbee40d9f6f624ce1105c4e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43015
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>