snapshot: Add Vm.freeze() and Vm.thaw() methods
When using vdsm images, we implement a snapshot using qcow format; we
add a new qcow image to the volume chain, with the current leaf as the
parent image, and then we ask libvirt to switch to the new image.
Libvirt is freezing the filesystems on the guest, switching the vm
drives to use the new volumes, and thawing the guest filesystems.

This flow cannot work for Ceph disks, which are using raw format. In
Ceph, snapshots are created on the server side, and we do not use volume
chains. After a snapshot it taken, the vm continue to write the same
image, and is not aware that the server took a snapshot of the disk.

To create consistent snapshot, we must freeze filesystems on the guest
before we take a Ceph snapshot, and we must thaw guest filesystems after
the snapshot was finished.

To support live snapshot when a vm is using Ceph disks, we will use this
flow:

- Engine tell vdsm to freeze guest's filesystems
- Engine create Ceph disk snapshot via Cinder
- Engine tell vdsm to take a snapshot (for other disks and memory
  snapshot)
- Engine tall vdsm to thaw guest's filesystems

This patch adds the new Vm methods invoking libvirt virDomainFSFreeze
and virDomainFSThaw.

Freezing or thawing filesystems can fail because of many reasons;
libvirt driver does not support freezing, guest agent is not installed
or not running or guest agent failure. We detect the first two cases, and
treat any other libvirt error as guest agent failure.

The errors are reported back to engine so it can handled them correctly.
The current policy implemented in vdsm is to log freezing failures
during snapshot, and retry the snapshot without freezing.  Future engine
versions can change this policy based on the error code returned from
new freezing APIs.

Change-Id: I8e4c54f461039de99823c3b80a10be0a960c4273
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43018
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Tested-by: Daniel Erez <derez@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>