hostdev: fix hostdev addressing
Host devices are special in a sense that in case of PCI, there are two
addresses present - one for the host address (in source element), and
one in the device body itself (how is the device seen on the host).

We have to make sure to separate and make sure which address do we put
where - the patch fixes an issue where the "guest" address was reported
as a "host" address.

Change-Id: I6d21555ec18d0e8f9766064755443785be3ef697
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43042
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>