hostdev: add tests for guest addressing
PCI host devices have two addresses: first one of them being the address
as seen on the host, the second one is how the device is seen in the
guest. We have to make sure that these aren't mixed - these tests are
here to catch possible address inconsistence.

Change-Id: I7bd20e1a2cb9846b5dc6d35315a89855b63a6ff4
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43093
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>