snapshot: Add vdsClient freeze() and thaw() APIs
These verbs are needed if you want to create a snapshot manually via the
shell on a vm containing a ceph disk. Note that this requires access to
cinder cli for creating the ceph snapshot, so this may not be very
useful. However, having these verbs in vdsClient make it easy to test
and debug these features.

Change-Id: I98872132767f3edd6d028cd9545e282e89f12777
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43098
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Tested-by: Daniel Erez <derez@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>