virt: Handle disconnects only if endpoints match
Until now we used to handle every disconnect event with no respect
of which remote endpoint caused it.
There are currently 3 different types of graphic events:
1. Connected event someone connected to the port
2. Initialized event - Someone had the right credentials to establish
   the connection and the session has started
3. Disconnected event - Someone disconnected from the port

Now any connected event can be triggered even with no credentials and
would eventually trigger the disconnected event due to being rejected.
This means that a host on the public internet which gets scanned for
opened ports and tried to intrude them would trigger the disconnect
handling for the affected VM.
This can cause that somone connected via Spice, VNC or even RDP with
the guest agent installed on the VM gets randomly a locked screen (default
behaviour).

This patch addresses the above mentioned issue by performing a comparison
of the remote endpoint that disconnected and remembering only initiated
endpoints. In case that multiple connections are established, like in case
of spice, the last initiated channel would be remembered and when that one
disconnects, it will cause the disconnect handling to proceed.

Due to some bug in the handling of graphic events in libvirt, SPICE
connections currently do not include the port. A fix for this has been
posted upstream and should eventually become available. Therefore, the
handling has been implemented to not ignore the disconnect event when
the clientPort value is an empty string.

Change-Id: Idfb3441407aa6e7bc31d37304bce0076e12bb1a3
Bug-Url: https://bugzilla.redhat.com/1203891
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43141
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI