vm: simplify Vm._getVmStatus method
Vm._getVmStatus returns a dict with just one key/value pair,
being the computed vm status (unsurprisingly).

This was done without a good reason, just to make the caller
look uniform. But more and more we need just the status value,
so the temporary dictionary is wasteful and redundant.

So, this patch simplifies the code to return indeed just the value.

Change-Id: I3f02efa1001aff5fbf01d127b333250500de9d54
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43267
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>