vm: event: emit event on domDependentInit
When Vm.domDependentInit() terminates, the VM
is lamost ready to go. If Engine queries the VM
on this state, on the happy path, we report either
POWERING_UP or UP depending also on the Guest Agent
state.

We should send an event to Engine ASAP, and toward
the end is indeed a good place. But we cannot lie,
ad assert the Vm is UP.

So, to be on the safe side, we send POWERING_UP
to Engine. The VM will be accessible through Console
and migrateable, so nothing is lost after this decision.

Change-Id: Ie379911a374a9ba3e82658592c295f7a9a0ac8a0
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43269
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>