StorageDomainManifest: Move clusterLock into Manifest
The clusterLock will be used by SDM verbs (as a per-domain lock) so it
needs to be moved into the Manifest so that it can be shared.

Moving the lock requires us to instantiate it in the
StorageDomainManifest constructor which requires some reordering of the
initialization steps.  This also affects the existing manifest_tests.
The DMDK_VERSION key must always be present in the metadata passed into
the StorageDomainManifest constructor since domainLock initialization
needs to check the value.  BlockStorageDomainManifest objects must now
always be created with a mocked LVM.

When running with an SPM, the SPM role is protected by the Master
Storage Domain's domain lock.  This lock is referred to as the
cluster lock since it covers all domains.  In SPM-less (SDM) flows, we
use each storage domain's lock as required.  For this reason, the
acquireClusterLock, releaseClusterLock, and inquireClusterLock APIs are
moved to acquireDomainLock, releaseDomainLock, and inquireDomainLock in
the StorageDomainManifest but their original names are preserved in the
StorageDomain.

Change-Id: I2c019a7fed6ba793d2b3a2459af221b1abbff2a3
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43549
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>