freeze: Avoid unneeded freeze during snapshot
When taking a snapshot of a vm with Ceph disk, Vm.snapshot() is called
after engine froze the vm. In this case we don't need freeze the guest
during the snapshot.

New engine versions will indicate the guest state using the new frozen
flag. When working with older engines or when not using Ceph disks,
behavior is unchanged.

Change-Id: I328470f700ab74daeed5375c0ade83628d16c387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43293
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>