virt: make vm events callback public
The Vm class has a few methods which are meant to
be called when libvirt reports events.
Examples:

Vm._onLibvirtLifecycleEvent
Vm._onWatchdogEvent
Vm._rtcUpdate

So they should be public, and they should be named such that
it is obvious they are meant to be callbacks.

This patch only renames without code changes.

Change-Id: Ib17737a3e4acb99bc67fae67919d71fe17ee771d
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43308
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>