Live Merge: Allow extension of non-leaf raw volumes
volume.extendSize() is currently prohibited for any non-leaf volume.
For a very specific live merge scenario we must permit extension of an
internal raw base volume.  Allow this usage and add a comment explaining
the reasoning.

The scenario:
 - User begins with a raw block disk.
 - User creates a snapshot.
 - User enlarges the disk (diskSizeExtend)
 - User performs live merge to remove the snapshot

In this case the base volume is too small to accommodate the data
from the child volume and an error is raised since libvirt cannot
enlarge a block device.  The solution is to require engine to call
extendVolumeSize on the base volume before requesting the live merge
operation.

Change-Id: I82a73356a863abe39d367cdc70a9b1e914f75aee
Signed-off-by: Adam Litke <alitke@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1232481
Reviewed-on: https://gerrit.ovirt.org/43407
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>