testlib: Fix permutations of decorated tests
Our expandPermutations infrastructure assume that generated tests use the
original test method, and break when using test decorators such as
@slowtestest or @stresstest.

The generated tests are registered incorrectly, overriding tests with the
same first permutation argument. Assuming we have the following test:

    @slowtest
    @permutations([[1, 2], [1, 4]])
    def test_permutations(self, a, b):
        ...

Instead of generating these tests:

    test_permutations(a=1, b=2)
    test_permutations(a=1, b=4)

We generate only one test, dropping the first permutation:

    test_permutations(kwargs=1)

This happens because decorated method signature is converted from:

    test_permutations(self, a, b)

To the generic form:

    test_permutations(*args, **kwargs)

When expanding permutations, we drop the first argument (args), assuming
it is self, and zipping the argument list with the rest of the argument
names (kwargs), dropping the second argument, and generating the wrong
argument string "kwargs=1".

The issue seems unfixable, since we need the original wrapped function
to get the original variables names, and we can have any number of
wrappers around it.

This patch fixes the issue in the same way nose and pytest handle this
issue, by generating positional argument string: "1, 2". This is not
nice as "a=1, b=2", but it works with decorated tests correctly.

So in the example above, we are generated these tests correctly:

    test_permutations(1, 2)
    test_permutations(1, 4)

Change-Id: I805885cb584f7ff015443682dd3bf52d0a44c74b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43433
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>