ssl: change default protocol
We used 'sslv23' but when Paddle (CVE-2014-3566) was found it
is recommended to switch to 'tlsv1'. On some of the OSes we
see issues that it is already sslv3 is disabled and users need
to switch to tls.

There is an issue when we switch to tls older engines (<= 3.0)
are not able to talk to vdsm anymore.


Change-Id: I40267cb07b19d444c7d85aba6d1160c27e8fe3a6
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Bug-Url: https://bugzilla.redhat.com/1229765
Reviewed-on: https://gerrit.ovirt.org/43457
Continuous-Integration: Jenkins CI
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/47832
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Yeela Kaplan <ykaplan@redhat.com>