vm: stats: do not send empty disks/networks
VDSM used to inconditionally add keys for disks and networks
in the stats value to be returned on Vm.getStats() API.

If stats from libvirt are available, those keys will be
overwritten with real data; otherwise, empty values are reported.

If we use the same code path to piggyback data on vm_status
events, this will confuse Engine and will make it believe
there are no disks nor networks.

Moreovero, in general, we should avoid to report empty values
if the emptiness carries no meaning (noteworthy counterexample:
vmJobs for live merge).

So this patch makes sure that we report either full data
for disks and networks, or nothing at all, avoiding empty values.

Change-Id: I30aac0079fec71b496d4db9536b65d8f16360d7a
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43464
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>