vm: stats: move vmstats translation code away
In vm class we have code to reformat and translate
the output of vmstats.produce() into the format that
Engine expects.

This is pure algorhytmic code which doesn't depend
at all on Vm class state, so it is wasy and safe to
extract it as pure function.

Additionally, this code is moved into vmstats module,
because of the (admittedly vague, but still stronger
than on Vm class) bonds it has with vmstats.produce()
function.

Change-Id: Ibc5d98c09defed392fbcd508be8d99116f2cbf91
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43466
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>