vm: events: do not override Down status
We want to reuse Vm._getVmStatus() in the events notification
flows.
However, in that method there is a race with the migration.SourceThread.
If the thread is still alive() (Vm.isMigrating() == True) while the
Vm._getVmStatus() runs, the Down status will not be reported;
MigrationSource will be reported instead.

Please note that this affects just and only the return value of
Vm._getVmStatus(). Vm.lastStatus is not affected at all.

To fix this misreport, we simply add "Down" status to the list
of statuses which should be reported directly without further checks.

The "Down" status is final, so there are no shortcomings
in sight and this change seems easy and safe.

Change-Id: I29c402b5e8299777f8292ad4815f84ec92ceba24
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43465
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI