network: wait for a bridge to appear before disabling IPv6 on it
A traceback such as

  File "/usr/share/vdsm/network/configurators/ifcfg.py", line 101, in configureBridge
    sysctl.disable_ipv6(bridge.name)
  File "/usr/lib/python2.7/site-packages/vdsm/sysctl.py", line 42, in disable_ipv6
    with open('/proc/sys/net/ipv6/conf/%s/disable_ipv6' % dev, 'w') as f:
IOError: [Errno 2] No such file or directory: u'/proc/sys/net/ipv6/conf/test-network/disable_ipv6'

shows that a bridge was not yet created when _ifup in ifcfg's configureBridge
already finished.

This patch makes configureBridge wait for the bridge for up to 1s. If it is
not created by then, a standard exception is triggered.

Change-Id: Ic5292e5a297ff50a0e82ba0781333e21fd9dd7f3
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43583
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>