vmstats: fix bulk stats misinterpretation (disk)
The current vmstats code is assuming that the bulk stats samples
are in a different format than they really are.
This mistake dates back to the lastest iterations of bulk stats
development, when various approaches to fecth and store bulk stats
samples were experimented, while looking for good performances.

In the end, we settled for the simplest approach, for the sake of
performance. Somehow, the translation patch was lost, most likely in
one bad rebase.

The effect of this mistake is that some disk stats
were lost, while others are still correctly reported.

This could explaain why it took so long to notice this failure: some
data was still present and correctly reported.
This patch recovers and reimplements the translation strategy,
and makes sure that we access the bulk stats samples using
the right keys.

Change-Id: I2d03dc914a4a9f729519ebd1365754e63833d300
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43682
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>