net: only backup ifcfg files if they are not already persisted.
5fc7217 introduced backing up of ifcfg files that were not owned by
vdsm before the current call. Now, after selective restoration was
introduced, when a network is restored because it is different
from it's persistence, it is edited. This is done by removing it
(and it's ifcfg files) an re-add it. The result is that ifcfgs get
backed up also because they were just deleted by the current call,
and are now re-added. We end up with file placeholders that make vdsm
ifdown them during restoration. This defeats the whole selective
restoration feature.
In order to preserve the original meaning of _backup, we now
compute all the possible ifcfg files that are part of unified
persistence and backup only the ones that are not.

Change-Id: I8d99efb8141cf7c6409050170c24714899c64be9
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1203422
Reviewed-on: https://gerrit.ovirt.org/43685
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>