vmstats: rename 'nic' function
The vmstats module exports a 'nic' function, which
produces the stats per-nic, in the format Engine expects.

The problem here is that 'nic' is a too nice name, so there
are other legitimate uses for it in other flows.

To reduce the possibility of name clash, we just rename
the function to the little less nice but little more revealing
'nic_traffic'.

This is also a little lie because 'nic_traffic' adds fields
which aren't strictly related to traffic proper. Future patch
will clean this further and remove the little lie.

Rename only.

Change-Id: I4b9290843c9dc7adc9709a66a0bb97238437bd37
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43694
Continuous-Integration: Jenkins CI
Reviewed-by: Marcin Mirecki <mmirecki@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>