clientIF: add logs during the recovery
The recovery flow is supposed to be sporadic, but it
is nevertheless very important.
The current flow is pretty opaque, and on big setups
it may take some time, leaving the admin to guess
what's going on.

This patch adds logs during the process, to make it
obvious what is happened, what is going on and what
is left to do.

Change-Id: I31dddf0a2bc760c5ad383ff6bfee9a72adc87c4f
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43770
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Dan Kenigsberg <danken@redhat.com>