vmstats: fix bulk stats misinterpretation (net)
The current vmstats code is assuming that the bulk stats samples
are in a different format than they really are.
This mistake dates back to the lastest iterations of bulk stats
development, when various approaches to fetch and store bulk stats
samples were experimented, while looking for good performances.

In the end, we settled for the simplest approach, for the sake of
performance. Somehow, the translation patch was lost, most likely in
one bad rebase.

The stats which are lost are the one which deals with multiple devices.
The bug is that we try to access the samples using wrong keys,
mistakenly assuming a nested namespace, while the samples are flat.

Change-Id: I57faae821394576aabc06dfaca2e1b4828a1339b
Bug-Url: https://bugzilla.redhat.com/1227793
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43679
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>