vmstats: handle known-missing stats
In some well-known and benign causes, libvirt will not
return the complete set of bulk stats.

It is dangerous to blindly ignore KeyErrors when some
fields are missing, because this can hide other bugs.
But it is safe to ignore them in the aforementioned cases.

This patch lays down the basic pieces to safely ignore
the errors, starting with the first loud and known case:
migration destination VM.

Backport-To: 3.6
Change-Id: I04ecbf98d36d576ede3c3cde36586245ea5e4218
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43777
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>