executor: Set worker system name
Previously we used to rename the workers using long name such as
"periodic-executor-worker-3". These names are too long for setting
thread system name (using pthread.setname()), and little too long in
general.

Since we managed worker ids in the Worker class, when creating multiple
executors, workers ids did not much the number of workers in the
executor.

This patch changes the naming scheme to "executor-name/<worker-id>",
similar to the way kernel threads are named.

For example, when creating executor named "periodic" with 4 workers, the
workers will use these names:

    periodic/0
    periodic/1
    periodic/2
    periodic/3

If we create another executor such as "jsonrpc", it will create these
worker names:

    jsonrpc/0
    jsonrpc/1
    jsonrpc/2
    jsonrpc/3
    jsonrpc/4
    jsonrpc/5
    jsonrpc/6
    jsonrpc/7

If the executor name is too long, the system thread name will be
truncated, as system thread names are limited to 15 characters.

Change-Id: Ia6a6b289daff2456213615355d6d6b51fd04b500
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43788
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>