executor: Fix race when putting tasks
When putting tasks into a TaskQueue, we used check the current number of
tasks in a thread-unsafe way.  We assume that the thread-unsafe check
can result in the worst case in few extra tasks in the task queue, which
is not very interesting.

However, if all workers are busy, and the task queue is almost full, it
is possible that two threads will pass the size check, and the queue
will contain more tasks then the maximum value. Once the queue size
exceeded the limit, we allow any number of additional tasks.

While this condition is unlikely, there is not reason to allow this
behavior. Now we check the queue size in a thread safe way and we don't
have to maintain the comment about the possible race.

Since we take the condition anyway on every put, I don't expect a
noticeable performance differences by taking the condition before the
size check.

Change-Id: Ie9ce28985e18629abb812e05752df90d05800c86
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43789
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>