executor: Fix thread-unsafe call
Deques support thread-safe, appends and pops from either side of the
deque, but deque.clear() is not documented as thread-safe, so we should
not assume it is.

Now we take the condition before invoking clear, and the comment about
thread-safety was fixed to match Python documentation.

Change-Id: Ia99d44698bc03bfcbac1a66f151f40220b8b4ef9
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43790
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1279950
Reviewed-on: https://gerrit.ovirt.org/48197
Tested-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Continuous-Integration: Francesco Romani <fromani@redhat.com>