xmlrpc: Fix the inheritance mess
Now that the xmlrpc server is not listening for connections, there is no
need to inherit from SimpleXMLRPCServer and its many super classes. The
xmlprc server is a very simple object that takes connected sockets from
a queue and starts a request handler for each socket in a new thread.
Most of the logic in the server is in the request handler class.

This patch implements this simple object, based on code from
SocketServer, BaseHTTPServer, and SimpleXMLRPCServer. This makes it easy
to understand and maintain, and remove dependencies on the fragile
Python server class hierarchy.

Change-Id: Ia912a465706b49f3f9146b2a74311fdd5a50f85e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43792
Continuous-Integration: Jenkins CI
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>