netinfo: Drop ifcfg reading from _getNetInfo
For backwards compatibility we need to report the bridge 'cfg' field
in the network attributes dictionary. Up until now, this was done by
re-reading the 'ifcfg' files of the bridge. After this change, we
will just be adding the 'cfg' field that we already generated for the
bridge device.

Note that 'BOOTPROTO' reporting is already part of _devinfo, so we
don't need that part.

Change-Id: I2a80251cc45792ba5511d22bce9451fb3b36689d
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Signed-off-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/35012
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI