gluster: Added VDSM verb to stop gluster related processes
Added a verb which stops the gluster related process like
brick processes, gsyncd process.
This needs to be done as part of a host moving to maintenance
mode. as data should not be available if a host is moved to
maintenence mode for some migration etc.

Change-Id: Id686e098b323eededcf1f89de331a1d524274995
Bug-URL: https://bugzilla.redhat.com/1189285
Signed-off-by: Shubhendu Tripathi <shtripat@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43821
Tested-by: Ramesh N <rnachimu@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Bala.FA <barumuga@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/48449
Continuous-Integration: Bala.FA <barumuga@redhat.com>