virt: Change state of the agent only after heartbeat processing
Until now we have been setting the guest agent state to up,
for every message we are receiving from the guest agent.

With the new events implementation this has the side effect,
that the data from the event has not been consumed yet but the
change will trigger the event being fired that the VM is UP.

This patch changes this old behaviour to only change the internal
state on heartbeat messages, and only after the data has been
consumed.

Change-Id: I541d3a4ad2930715d38c8908e2c5d6543ad72037
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43859
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Tested-by: Omer Frenkel <ofrenkel@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>