net: configurators: persist custom bond option
Until now, custom bond property was exposed only to
before_network_setup hook and then dropped (so it was not
passed to configurators and persisted). But this is not how
custom *network* property behaves. We want these two custom
properties to be symmetric - they have to be persisted
(if unified persistence is used). Unlike custom network
properties (which are known as 'special' by Engine), custom bond
properties have to be reported in netinfo. Last but not least,
we need this custom property to be able to tag bonding as
'OVS Bond' so it will be handled by OVS hook, not by default
configurator.

This patch moves removal of custom property down to
configurators. Thanks to it, custom property is persisted,
but not set as a bond option.

Custom property is not dropped only in configurators, but in
models.py:Bond:areOptionsApplied as well - we don't want
configurator to reconfigure bond when there is a custom
property defined while its value is not used by configurator.

Change-Id: Ic4f1564e19904cc1d53bc6d1cf732ca35375332e
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/43891
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI